Implemented pagination and sorting for user table

+++ src/main/java/com/animalmanagementsystem/shelter/services/impl/UserServiceImpl.java
@@ -14,6 +14,10 @@ import jakarta.persistence.criteria.CriteriaQuery;
 import jakarta.persistence.criteria.Predicate;
 import jakarta.persistence.criteria.Root;
 import jakarta.transaction.Transactional;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
@@ -72,9 +76,11 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public List<UserDto> getAllUsers() {
-        List<UserEntity> userEntities = userRepository.findAll();
-        return userEntities.stream().map(userMapper::mapEntityToDto).toList();
+    public List<UserDto> getAllUsers(int pageNo, int pageSize) {
+        Pageable pageable = PageRequest.of(pageNo, pageSize, Sort.by(Sort.Direction.ASC, "email"));
+        Page<UserEntity> userEntities = userRepository.findAll(pageable);
+        List<UserEntity> list = userEntities.getContent();
+        return list.stream().map(userMapper::mapEntityToDto).toList();
     }
 
     @Override
@@ -99,7 +105,9 @@ public class UserServiceImpl implements UserService {
     public void deleteUser(Long id) {
         Optional<UserEntity> optionalUserEntity = userRepository.findById(id);
 
-        optionalUserEntity.orElseThrow(() -> new UserNotFoundException(id));
+        if (optionalUserEntity.isEmpty()) {
+            throw new UserNotFoundException(id);
+        }
 
         userRepository.deleteById(id);
     }
